requirements.txt
torch>=1.11.0
transformers>=4.17.0
numpy>=1.22.0
pandas>=1.4.0
scikit-learn>=1.0.0
xgboost>=1.6.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
pyyaml>=6.0
cma>=3.3.0
hdbscan>=0.8.0
umap-learn>=0.5.0
scipy>=1.9.0
